TreeQueryNodeServerExecutable/src/main/java/org/treequery/grpc/hostutils/HostInfo.java
package org.treequery.grpc.hostutils;

import com.google.api.client.util.Lists;
import lombok.SneakyThrows;
import lombok.extern.slf4j.Slf4j;

import java.net.InetAddress;
import java.net.NetworkInterface;
import java.net.SocketException;
import java.util.Enumeration;
import java.util.List;

@Slf4j
public class HostInfo {

    @SneakyThrows
    public static List<String> getMyAddress() {
        String ip;
        List<String> ipList = Lists.newArrayList();

        Enumeration<NetworkInterface> interfaces = NetworkInterface.getNetworkInterfaces();
        while (interfaces.hasMoreElements()) {
            NetworkInterface iface = interfaces.nextElement();
            // filters out 127.0.0.1 and inactive interfaces
            if (iface.isLoopback() || !iface.isUp())
                continue;

            Enumeration<InetAddress> addresses = iface.getInetAddresses();
            while (addresses.hasMoreElements()) {
                InetAddress addr = addresses.nextElement();
                ip = addr.getHostAddress();
                ipList.add((ip));
                log.info(iface.getDisplayName() + " " + ip);
            }
        }
        return ipList;
    }

}
